+++ transport/api/Cargo.toml
@@ -39,7 +39,6 @@ async-std = { workspace = true, optional = true }
 async-trait = { workspace = true }
 async-lock = { workspace = true }
 futures = { workspace = true }
-futures-concurrency = "7.6.0"
 lazy_static = { workspace = true, optional = true }
 libp2p = { workspace = true, features = ["request-response"] }
 proc-macro-regex = "~1.1.0"
