+++ transport/p2p/Cargo.toml
@@ -18,6 +18,7 @@ async-std = { workspace = true, optional = true}
 async-channel = { workspace = true }
 futures = { workspace = true }
 futures-lite = { workspace = true }
+futures-concurrency = { workspace = true }
 libp2p = { workspace = true, features = ["noise", "request-response", "cbor", "macros", "tcp", "quic", "dns", "yamux"] }
 libp2p-mplex = { workspace = true }
 multiaddr = { workspace = true }
@@ -30,6 +31,7 @@ void = "1"          # needed for Behavior implementations from libp2p
 core-network = { workspace = true }
 core-protocol = { workspace = true }
 hopr-internal-types = { workspace = true }
+hopr-crypto-types = { workspace = true }
 
 [dev-dependencies]
 async-std = { workspace = true }
