+++ transport/p2p/src/swarm.rs
@@ -1,5 +1,11 @@
-use std::{net::Ipv4Addr, num::NonZeroU8};
-use tracing::{error, info};
+use futures::{channel::mpsc::UnboundedSender, pin_mut, select, StreamExt};
+use futures_concurrency::stream::Merge;
+use libp2p::{request_response::OutboundRequestId, PeerId};
+use std::{
+    collections::{HashMap, HashSet},
+    num::NonZeroU8,
+};
+use tracing::{debug, error, info, trace, warn};
 
 // #[cfg(any(feature = "runtime-async-std", test))]
 // use async_std::task::spawn;
@@ -7,70 +13,34 @@ use tracing::{error, info};
 // #[cfg(all(feature = "runtime-tokio", not(test)))]
 // use tokio::task::spawn;
 
+use core_network::{messaging::ControlMessage, network::NetworkTriggeredEvent, ping::PingQueryReplier};
+use core_protocol::{
+    ack::processor::{AckProcessed, AckResult, AcknowledgementInteraction},
+    config::ProtocolConfig,
+    msg::processor::{MsgProcessed, PacketInteraction},
+    ticket_aggregation::processor::{
+        TicketAggregationFinalizer, TicketAggregationInteraction, TicketAggregationProcessed,
+    },
+};
+use hopr_internal_types::prelude::*;
+
 use crate::{
     constants,
     errors::Result,
-    multiaddrs::{Multiaddr, Protocol},
-    HoprNetworkBehavior,
+    libp2p::{request_response::ResponseChannel, swarm::SwarmEvent},
+    multiaddrs::{replace_transport_with_unspecified, resolve_dns_if_any, Multiaddr},
+    HoprNetworkBehavior, HoprNetworkBehaviorEvent, PeerTransportEvent, Ping, Pong, TransportOutput,
 };
-use core_protocol::config::ProtocolConfig;
 
-use std::net::ToSocketAddrs;
+#[cfg(all(feature = "prometheus", not(test)))]
+use hopr_metrics::metrics::SimpleGauge;
 
-/// Replaces the IPv4 and IPv6 from the network layer with a unspecified interface in any multiaddress.
-fn replace_transport_with_unspecified(ma: &Multiaddr) -> Result<Multiaddr> {
-    let mut out = Multiaddr::empty();
-
-    for proto in ma.iter() {
-        match proto {
-            Protocol::Ip4(_) => out.push(std::net::IpAddr::V4(Ipv4Addr::UNSPECIFIED).into()),
-            Protocol::Ip6(_) => out.push(std::net::IpAddr::V6(std::net::Ipv6Addr::UNSPECIFIED).into()),
-            _ => out.push(proto),
-        }
-    }
-
-    Ok(out)
-}
-
-/// Resolves the DNS parts of a multiaddress and replaces it with the resolved IP address.
-fn resolve_dns_if_any(ma: &Multiaddr) -> Result<Multiaddr> {
-    let mut out = Multiaddr::empty();
-
-    for proto in ma.iter() {
-        match proto {
-            Protocol::Dns4(domain) => {
-                let ip = format!("{domain}:443") // dummy port, irrevelant at this point
-                    .to_socket_addrs()
-                    .map_err(|e| crate::errors::P2PError::Logic(e.to_string()))?
-                    .filter(|sa| sa.is_ipv4())
-                    .collect::<Vec<_>>()
-                    .first()
-                    .ok_or(crate::errors::P2PError::Logic(format!(
-                        "Failed to resolve {domain} to an IPv4 address. Does the DNS entry has an A record?"
-                    )))?
-                    .ip();
-
-                out.push(ip.into())
-            }
-            Protocol::Dns6(domain) => {
-                let ip = format!("{domain}:443") // dummy port, irrevelant at this point
-                    .to_socket_addrs()
-                    .map_err(|e| crate::errors::P2PError::Logic(e.to_string()))?
-                    .filter(|sa| sa.is_ipv6())
-                    .collect::<Vec<_>>()
-                    .first()
-                    .ok_or(crate::errors::P2PError::Logic(format!(
-                        "Failed to resolve {domain} to an IPv6 address. Does the DNS entry has an AAAA record?"
-                    )))?
-                    .ip();
-
-                out.push(ip.into())
-            }
-            _ => out.push(proto),
-        }
-    }
-
-    Ok(out)
+#[cfg(all(feature = "prometheus", not(test)))]
+lazy_static::lazy_static! {
+    static ref METRIC_TRANSPORT_P2P_OPEN_CONNECTION_COUNT: SimpleGauge = SimpleGauge::new(
+        "hopr_transport_p2p_opened_connection_count",
+        "Number of currently open connections"
+    ).unwrap();
 }
 
 /// Build objects comprising the p2p network.
@@ -187,10 +157,536 @@ impl HoprSwarm {
 
         Self { swarm }
     }
+
+    pub fn with_processors(
+        self,
+        network_update_input: futures::channel::mpsc::Receiver<NetworkTriggeredEvent>,
+        indexer_update_input: futures::channel::mpsc::UnboundedReceiver<PeerTransportEvent>,
+        ack_interactions: AcknowledgementInteraction,
+        pkt_interactions: PacketInteraction,
+        ticket_aggregation_interactions: TicketAggregationInteraction<
+            TicketAggregationResponseType,
+            TicketAggregationRequestType,
+        >,
+        heartbeat_requests: futures::channel::mpsc::UnboundedReceiver<(PeerId, PingQueryReplier)>,
+    ) -> HoprSwarmWithProcessors {
+        HoprSwarmWithProcessors {
+            swarm: self,
+            network_update_input,
+            indexer_update_input,
+            ack_interactions,
+            pkt_interactions,
+            ticket_aggregation_interactions,
+            heartbeat_requests,
+        }
+    }
 }
 
 impl From<HoprSwarm> for libp2p::Swarm<HoprNetworkBehavior> {
     fn from(value: HoprSwarm) -> Self {
         value.swarm
+    }
+}
+
+/// Composition of all inputs allowing to produce a single stream of
+/// input events passed into the swarm processing logic.
+#[derive(Debug)]
+pub enum Inputs {
+    Heartbeat((PeerId, PingQueryReplier)),
+    NetworkUpdate(NetworkTriggeredEvent),
+    Message(MsgProcessed),
+    TicketAggregation(
+        TicketAggregationProcessed<ResponseChannel<std::result::Result<Ticket, String>>, OutboundRequestId>,
+    ),
+    Acknowledgement(AckProcessed),
+    Indexer(PeerTransportEvent),
+}
+
+impl From<(PeerId, PingQueryReplier)> for Inputs {
+    fn from(value: (PeerId, PingQueryReplier)) -> Self {
+        Self::Heartbeat(value)
+    }
+}
+
+impl From<NetworkTriggeredEvent> for Inputs {
+    fn from(value: NetworkTriggeredEvent) -> Self {
+        Self::NetworkUpdate(value)
+    }
+}
+
+impl From<AckProcessed> for Inputs {
+    fn from(value: AckProcessed) -> Self {
+        Self::Acknowledgement(value)
+    }
+}
+
+impl From<MsgProcessed> for Inputs {
+    fn from(value: MsgProcessed) -> Self {
+        Self::Message(value)
+    }
+}
+
+impl From<TicketAggregationProcessed<TicketAggregationResponseType, TicketAggregationRequestType>> for Inputs {
+    fn from(value: TicketAggregationProcessed<TicketAggregationResponseType, TicketAggregationRequestType>) -> Self {
+        Self::TicketAggregation(value)
+    }
+}
+
+impl From<PeerTransportEvent> for Inputs {
+    fn from(value: PeerTransportEvent) -> Self {
+        Self::Indexer(value)
+    }
+}
+
+use hopr_internal_types::legacy;
+
+pub type TicketAggregationRequestType = OutboundRequestId;
+pub type TicketAggregationResponseType = ResponseChannel<std::result::Result<Ticket, String>>;
+
+pub struct HoprSwarmWithProcessors {
+    swarm: HoprSwarm,
+    network_update_input: futures::channel::mpsc::Receiver<NetworkTriggeredEvent>,
+    indexer_update_input: futures::channel::mpsc::UnboundedReceiver<PeerTransportEvent>,
+    ack_interactions: AcknowledgementInteraction,
+    pkt_interactions: PacketInteraction,
+    ticket_aggregation_interactions:
+        TicketAggregationInteraction<TicketAggregationResponseType, TicketAggregationRequestType>,
+    heartbeat_requests: futures::channel::mpsc::UnboundedReceiver<(PeerId, PingQueryReplier)>,
+}
+
+impl std::fmt::Debug for HoprSwarmWithProcessors {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("SwarmEventLoop").finish()
+    }
+}
+
+impl HoprSwarmWithProcessors {
+    /// Main p2p loop that instantiates a new libp2p::Swarm instance and sets up listening and reacting pipelines
+    /// running in a neverending loop future.
+    ///
+    /// The function represents the entirety of the business logic of the hopr daemon related to core operations.
+    ///
+    /// This future can only be resolved by an unrecoverable error or a panic.
+    pub async fn run(
+        self,
+        version: String,
+        on_transport_output: UnboundedSender<TransportOutput>,
+        on_acknowledged_ticket: UnboundedSender<AcknowledgedTicket>,
+    ) {
+        let mut swarm: libp2p::Swarm<HoprNetworkBehavior> = self.swarm.into();
+
+        let mut ack_writer = self.ack_interactions.writer();
+        let mut pkt_writer = self.pkt_interactions.writer();
+        let mut aggregation_writer = self.ticket_aggregation_interactions.writer();
+
+        // NOTE: an improvement would be a forgetting cache for the active requests
+        let mut active_pings: HashMap<libp2p::request_response::OutboundRequestId, PingQueryReplier> = HashMap::new();
+        let mut active_aggregation_requests: HashMap<
+            libp2p::request_response::OutboundRequestId,
+            TicketAggregationFinalizer,
+        > = HashMap::new();
+
+        let mut allowed_peers: HashSet<PeerId> = HashSet::new();
+
+        let inputs = (
+            self.heartbeat_requests.map(Inputs::Heartbeat),
+            self.network_update_input.map(Inputs::NetworkUpdate),
+            self.ack_interactions.map(Inputs::Acknowledgement),
+            self.pkt_interactions.map(Inputs::Message),
+            self.ticket_aggregation_interactions.map(Inputs::TicketAggregation),
+            self.indexer_update_input.map(Inputs::Indexer),
+        )
+            .merge()
+            .fuse();
+
+        pin_mut!(inputs);
+
+        loop {
+            select! {
+                input = inputs.select_next_some() => match input {
+                    Inputs::Heartbeat((peer, replier)) => {
+                        trace!("transport input - heartbeat - executing ping to peer '{peer}'");
+                        let req_id = swarm.behaviour_mut().heartbeat.send_request(&peer, Ping(replier.challenge()));
+                        active_pings.insert(req_id, replier);
+                    },
+                    Inputs::NetworkUpdate(event) => match event {
+                        NetworkTriggeredEvent::CloseConnection(peer) => {
+                            debug!("transport input - network event - closing connection to '{peer}' (reason: low ping connection quality)");
+                            if swarm.is_connected(&peer) {
+                                let _ = swarm.disconnect_peer_id(peer);
+                            }
+                        },
+                        NetworkTriggeredEvent::UpdateQuality(_, _) => {}
+                    },
+                    Inputs::Acknowledgement(task) => match task {
+                        AckProcessed::Receive(peer, reply) => {
+                            debug!("transport input - ack - received an acknowledgement from '{peer}'");
+                            if let Ok(reply) = reply {
+                                match reply {
+                                    AckResult::Sender(half_key_challenge) => {
+                                        if let Err(e) = on_transport_output.unbounded_send(TransportOutput::Sent(half_key_challenge)) {
+                                            error!("transport input - ack - failed to emit received acknowledgement: {e}")
+                                        }
+                                    },
+                                    AckResult::RelayerWinning(acknowledged_ticket) => {
+                                        if let Err(e) = on_acknowledged_ticket.unbounded_send(acknowledged_ticket) {
+                                            error!("transport input - ack -failed to emit acknowledged ticket: {e}");
+                                        }
+                                    }
+                                    AckResult::RelayerLosing => {}
+                                }
+                            }
+                        },
+                        AckProcessed::Send(peer, ack) => {
+                            trace!("transport input - ack - sending an acknowledgement to '{peer}'");
+                            let _req_id = swarm.behaviour_mut().ack.send_request(&peer, ack);
+                        }
+                    }
+                    Inputs::Message(task) => match task {
+                        MsgProcessed::Receive(peer, data, ack) => {
+                            debug!("transport input - msg - received packet from '{peer}'");
+                            if let Err(e) = on_transport_output.unbounded_send(TransportOutput::Received(data)) {
+                                error!("transport input - msg - failed to store a received message in the inbox: {}", e);
+                            }
+
+                            if let Err(e) = ack_writer.send_acknowledgement(peer, ack) {
+                                error!("transport input - msg - failed to acknowledge the received final packet: {e}");
+                            }
+                        },
+                        MsgProcessed::Send(peer, octets) => {
+                            debug!("transport input - msg - sending packet as source to '{peer}'");
+                            let _request_id = swarm.behaviour_mut().msg.send_request(&peer, octets);
+                        },
+                        MsgProcessed::Forward(peer, octets, previous_peer, ack) => {
+                            debug!("transport input - msg - forwarding packet from '{previous_peer}' to '{peer}'");
+                            let _request_id = swarm.behaviour_mut().msg.send_request(&peer, octets);
+                            if let Err(e) = ack_writer.send_acknowledgement(previous_peer, ack) {
+                                error!("transport input - msg - failed to acknowledge relayed packet: {e}");
+                            }
+                        }
+                    },
+                    Inputs::TicketAggregation(task) => match task {
+                        TicketAggregationProcessed::Send(peer, acked_tickets, finalizer) => {
+                            debug!("transport input - ticket aggregation - send request to '{peer}' to aggregate {} tickets", acked_tickets.len());
+                            let req_id = swarm.behaviour_mut().ticket_aggregation.send_request(&peer, acked_tickets);
+                            active_aggregation_requests.insert(req_id, finalizer);
+                        },
+                        TicketAggregationProcessed::Reply(peer, ticket, response) => {
+                            debug!("transport input - ticket aggregation - responding to request by '{peer}'");
+                            if swarm.behaviour_mut().ticket_aggregation.send_response(response, ticket).is_err() {
+                                error!("transport input - ticket aggregation - failed to send response to '{peer}'");
+                            }
+                        },
+                        TicketAggregationProcessed::Receive(peer, acked_ticket, request) => {
+                            if let Err(e) = on_acknowledged_ticket.unbounded_send(acked_ticket) {
+                                error!("transport input - ticket aggregation - failed to emit acknowledged aggregated ticket to '{peer}': {e}");
+                            }
+
+                            match active_aggregation_requests.remove(&request) {
+                                Some(finalizer) => finalizer.finalize(),
+                                None => {
+                                    warn!("transport input - ticket aggregation - response already handled")
+                                }
+                            }
+                        }
+                    },
+                    Inputs::Indexer(task) => match task {
+                        PeerTransportEvent::Allow(peer) => {
+                            debug!("transport input - indexer - allowing '{peer}'");
+                            let _ = allowed_peers.insert(peer);
+                        }
+                        PeerTransportEvent::Ban(peer) => {
+                            debug!("transport input - indexer - banning '{peer}'");
+                            allowed_peers.remove(&peer);
+
+                            if swarm.is_connected(&peer) {
+                                match swarm.disconnect_peer_id(peer) {
+                                    Ok(_) => debug!("Peer '{peer}' disconnected on network registry update"),
+                                    Err(e) => error!("Failed to disconnect peer '{peer}' on network registry update: {:?}", e)
+                                }
+                            }
+                        },
+                        PeerTransportEvent::Announce(peer, multiaddresses) => {
+                            if &peer != swarm.local_peer_id() {
+                                trace!("transport input - indexer - processing announcement for '{peer}' with addresses: '{multiaddresses:?}'");
+                                for multiaddress in multiaddresses.iter() {
+                                    if !swarm.is_connected(&peer) {
+                                        match swarm.dial(multiaddress.clone()) {
+                                            Ok(_) => {
+                                                debug!("transport input - indexer - storing '{multiaddress}' as valid for '{peer}'");
+                                                swarm.behaviour_mut().heartbeat.add_address(&peer, multiaddress.clone());
+                                                swarm.behaviour_mut().msg.add_address(&peer, multiaddress.clone());
+                                                swarm.behaviour_mut().ack.add_address(&peer, multiaddress.clone());
+                                                swarm.behaviour_mut().ticket_aggregation.add_address(&peer, multiaddress.clone());
+                                            },
+                                            Err(e) => {
+                                                warn!("transport input - indexer - failed to dial an announced peer '{peer}': {e}, ignoring the address '{multiaddress}'");
+                                            }
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                },
+                event = swarm.select_next_some() => match event {
+                    // ---------------
+                    // msg protocol
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::Message(libp2p::request_response::Event::<Box<[u8]>, ()>::Message {
+                        peer,
+                        message:
+                        libp2p::request_response::Message::<Box<[u8]>, ()>::Request {
+                            request_id, request, channel
+                        },
+                    })) => {
+                        debug!("transport protocol - p2p - msg - received a message from {peer}");
+
+                        if let Err(e) = pkt_writer.receive_packet(request, peer) {
+                            error!("transport protocol - p2p - msg - failed to process a message from '{peer}': {e} (#{request_id})");
+                        };
+
+                        if swarm.behaviour_mut().msg.send_response(channel, ()).is_err() {
+                            error!("transport protocol - p2p - msg - failed to send a response to '{peer}', likely a timeout");
+                        };
+                    },
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::Message(libp2p::request_response::Event::<Box<[u8]>, ()>::Message {
+                        peer,
+                        message:
+                        libp2p::request_response::Message::<Box<[u8]>, ()>::Response {
+                            request_id, ..
+                        },
+                    })) => {
+                        trace!("transport protocol - p2p - msg - received a response for sending message with id {request_id} from '{peer}'");
+                    },
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::Message(libp2p::request_response::Event::<Box<[u8]>, ()>::OutboundFailure {
+                        peer, error, request_id
+                    })) => {
+                        error!("transport protocol - p2p - msg - failed to send a message (#{}) to peer {} with an error: {}", request_id, peer, error);
+                    },
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::Message(libp2p::request_response::Event::<Box<[u8]>, ()>::InboundFailure {..}))
+                    | SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::Message(libp2p::request_response::Event::<Box<[u8]>, ()>::ResponseSent {..})) => {
+                    },
+                    // ---------------
+                    // ack protocol
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::Acknowledgement(libp2p::request_response::Event::<Acknowledgement,()>::Message {
+                        peer,
+                        message:
+                        libp2p::request_response::Message::<Acknowledgement,()>::Request {
+                            request_id, request, channel
+                        },
+                    })) => {
+                        debug!("transport protocol - p2p - ack - received an acknowledgment from '{peer}'");
+
+                        if let Err(e) = ack_writer.receive_acknowledgement(peer, request) {
+                            error!("transport protocol - p2p - ack - failed to process an acknowledgement from '{peer}': {e} (#{request_id})");
+                        };
+
+                        if swarm.behaviour_mut().ack.send_response(channel, ()).is_err() {
+                            error!("transport protocol - p2p - ack - failed to send a response to '{peer}', likely a timeout");
+                        };
+                    },
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::Acknowledgement(libp2p::request_response::Event::<Acknowledgement,()>::Message {
+                        peer,
+                        message:
+                        libp2p::request_response::Message::<Acknowledgement,()>::Response {
+                            request_id, ..
+                        },
+                    })) => {
+                        trace!("transport protocol - p2p - ack - received a response for sending message with id {} from {}", request_id, peer);
+                    },
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::Acknowledgement(libp2p::request_response::Event::<Acknowledgement,()>::OutboundFailure {
+                        peer, error, ..
+                    })) => {
+                        error!("transport protocol - p2p - ack - failed to send an acknowledgement to '{peer}': {error}");
+                    },
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::Acknowledgement(libp2p::request_response::Event::<Acknowledgement,()>::InboundFailure {..}))
+                    | SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::Acknowledgement(libp2p::request_response::Event::<Acknowledgement,()>::ResponseSent {..})) => {
+                        // debug!("Discarded messages not relevant for the protocol!");
+                    },
+                    // --------------
+                    // ticket aggregation protocol
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::TicketAggregation(libp2p::request_response::Event::<Vec<legacy::AcknowledgedTicket>, std::result::Result<Ticket,String>>::Message {
+                        peer,
+                        message:
+                        libp2p::request_response::Message::<Vec<legacy::AcknowledgedTicket>, std::result::Result<Ticket,String>>::Request {
+                            request_id, request, channel
+                        },
+                    })) => {
+                        debug!("transport protocol - p2p - ticket aggregation - received an aggregation request {request_id} from '{peer}'");
+                        let request = request.into_iter().map(TransferableWinningTicket::from).collect::<Vec<_>>();
+                        if let Err(e) = aggregation_writer.receive_aggregation_request(peer, request, channel) {
+                            error!("transport protocol - p2p - ticket aggregation - failed to aggregate tickets for '{peer}': {e}");
+                        }
+                    },
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::TicketAggregation(libp2p::request_response::Event::<Vec<legacy::AcknowledgedTicket>, std::result::Result<Ticket,String>>::Message {
+                        peer,
+                        message:
+                        libp2p::request_response::Message::<Vec<legacy::AcknowledgedTicket>, std::result::Result<Ticket,String>>::Response {
+                            request_id, response
+                        },
+                    })) => {
+                        if let Err(e) = aggregation_writer.receive_ticket(peer, response, request_id) {
+                            error!("transport protocol - p2p - ticket aggregation - error while handling aggregated ticket from '{peer}': {e}");
+                        }
+                    },
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::TicketAggregation(libp2p::request_response::Event::<Vec<legacy::AcknowledgedTicket>, std::result::Result<Ticket,String>>::OutboundFailure {
+                        peer, request_id, error,
+                    })) => {
+                        error!("transport protocol - p2p - ticket aggregation - failed to send an aggergation request #{request_id} to {peer}: {error}");
+                    },
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::TicketAggregation(libp2p::request_response::Event::<Vec<legacy::AcknowledgedTicket>, std::result::Result<Ticket,String>>::InboundFailure {
+                        peer, request_id, error})) => {
+                        warn!("transport protocol - p2p - ticket aggregation - encountered inbound failure for '{peer}' (#{request_id}): {error}")
+                    }
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::TicketAggregation(libp2p::request_response::Event::<Vec<legacy::AcknowledgedTicket>, std::result::Result<Ticket,String>>::ResponseSent {..})) => {
+                    },
+                    // --------------
+                    // heartbeat protocol
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::Heartbeat(libp2p::request_response::Event::<Ping,Pong>::Message {
+                        peer,
+                        message:
+                        libp2p::request_response::Message::<Ping,Pong>::Request {
+                            request_id, request, channel
+                        },
+                    })) => {
+                        trace!("transport protocol - p2p - heartbeat - received a Ping request from '{peer}' (#{request_id})");
+                        if let Ok(challenge_response) = ControlMessage::generate_pong_response(&request.0)
+                        {
+                            if swarm.behaviour_mut().heartbeat.send_response(channel, Pong(challenge_response, version.clone())).is_err() {
+                                error!("transport protocol - p2p - heartbeat - failed to reply to ping request");
+                            };
+                        }
+                    },
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::Heartbeat(libp2p::request_response::Event::<Ping,Pong>::Message {
+                        peer,
+                        message:
+                        libp2p::request_response::Message::<Ping,Pong>::Response {
+                            request_id, response
+                        },
+                    })) => {
+                        if let Some(replier) = active_pings.remove(&request_id) {
+                            trace!("transport protocol - p2p - heartbeat - processing manual ping response from '{peer}'");
+                            replier.notify(response.0, response.1)
+                        } else {
+                            debug!("transport protocol - p2p - heartbeat - failed to find heartbeat replier for '{peer}'");
+                        }
+                    },
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::Heartbeat(libp2p::request_response::Event::<Ping,Pong>::OutboundFailure {
+                        request_id, peer, error,
+                    })) => {
+                        trace!("transport protocol - p2p - heartbeat - encountered outbound failure for '{peer}': {error}");
+                        active_pings.remove(&request_id);
+                    },
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::Heartbeat(libp2p::request_response::Event::<Ping,Pong>::InboundFailure {
+                        peer, error, ..     // request_id
+                    })) => {
+                        warn!("transport protocol - p2p - heartbeat - inbound failure for peer '{peer}': {error}")
+                    }
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::Heartbeat(libp2p::request_response::Event::<Ping,Pong>::ResponseSent {..})) => {},
+                    SwarmEvent::Behaviour(HoprNetworkBehaviorEvent::KeepAlive(_)) => {}
+                    SwarmEvent::ConnectionEstablished {
+                        peer_id,
+                        connection_id,
+                        ..
+                        // endpoint,
+                        // num_established,
+                        // concurrent_dial_errors,
+                        // established_in,
+                    } => {
+                        debug!("transport - p2p - connection ({connection_id}) established with {peer_id}");
+
+                        #[cfg(all(feature = "prometheus", not(test)))]
+                        {
+                            METRIC_TRANSPORT_P2P_OPEN_CONNECTION_COUNT.increment(1.0);
+                        }
+
+                        if !allowed_peers.contains(&peer_id) {
+                            info!("transport - p2p - DISCONNECTING '{peer_id}': not allowed in the network registry)");
+                            let _ = swarm.disconnect_peer_id(peer_id);
+                        }
+                    },
+                    SwarmEvent::ConnectionClosed {
+                        peer_id,
+                        cause,
+                        ..
+                        // connection_id,
+                        // endpoint,
+                        // num_established,
+                    } => {
+                        debug!("transport - p2p - connection closed for peer '{peer_id}': {cause:?}");
+
+                        #[cfg(all(feature = "prometheus", not(test)))]
+                        {
+                            METRIC_TRANSPORT_P2P_OPEN_CONNECTION_COUNT.decrement(1.0);
+                        }
+                    },
+                    SwarmEvent::IncomingConnection {
+                        connection_id,
+                        local_addr,
+                        send_back_addr,
+                    } => {
+                        debug!("transport - p2p - incoming connection at {local_addr} from {send_back_addr} ({connection_id:?})");
+                    },
+                    SwarmEvent::IncomingConnectionError {
+                        local_addr,
+                        error,
+                        ..
+                        // connection_id,
+                        // send_back_addr,
+                    } => {
+                        error!("transport - p2p - incoming connection error on {local_addr}: {error}")
+                    },
+                    SwarmEvent::OutgoingConnectionError {
+                        connection_id,
+                        error,
+                        peer_id
+                    } => {
+                        error!("transport - p2p - outgoing connection error for peer '{peer_id:?}' ({connection_id:?}): {error}")
+                    },
+                    SwarmEvent::NewListenAddr {
+                        listener_id,
+                        ..
+                        // address,
+                    } => {
+                        debug!("transport - p2p - new listen address on {listener_id:?}")
+                    },
+                    SwarmEvent::ExpiredListenAddr {
+                        listener_id,
+                        ..
+                        // address,
+                    } => {
+                        debug!("transport - p2p - expired listen address on {listener_id:?}")
+                    },
+                    SwarmEvent::ListenerClosed {
+                        listener_id,
+                        ..
+                        // addresses,
+                        // reason,
+                    } => {
+                        debug!("transport - p2p - listener closed on {listener_id:?}", )
+                    },
+                    SwarmEvent::ListenerError {
+                        listener_id,
+                        error,
+                    } => {
+                        debug!("transport - p2p - listener error for {listener_id:?}: {error}")
+                    },
+                    SwarmEvent::Dialing {
+                        peer_id,
+                        connection_id,
+                    } => {
+                        if let Some(peer_id) = peer_id {
+                            if !allowed_peers.contains(&peer_id) {
+                                info!("transport - p2p - dialing '{peer_id}': not allowed in the network registry)");
+                                let _ = swarm.disconnect_peer_id(peer_id);
+                            } else {
+                                debug!("transport - p2p - dialing peer {peer_id:?} ({connection_id:?}")
+                            }
+                        }
+                    },
+                    _ => error!("transport - p2p - unimplemented message type in p2p processing chain encountered")
+                }
+            }
+        }
     }
 }
