+++ transport/protocol/src/ack/processor.rs
@@ -140,37 +140,30 @@ impl AcknowledgementInteraction {
                             if ack.validate(&remote_pk) {
                                 match processor.handle_acknowledgement(ack).await {
                                     Ok(reply) => {
+                                        #[cfg(all(feature = "prometheus", not(test)))]
                                         match &reply {
                                             AckResult::Sender(_) => {
-                                                #[cfg(all(feature = "prometheus", not(test)))]
                                                 METRIC_RECEIVED_ACKS.increment(&["true"]);
                                             }
                                             AckResult::RelayerWinning(_) => {
-                                                #[cfg(all(feature = "prometheus", not(test)))]
-                                                {
-                                                    METRIC_RECEIVED_ACKS.increment(&["true"]);
-                                                    METRIC_TICKETS_COUNT.increment(&["winning"]);
-                                                }
+                                                METRIC_RECEIVED_ACKS.increment(&["true"]);
+                                                METRIC_TICKETS_COUNT.increment(&["winning"]);
                                             }
                                             AckResult::RelayerLosing => {
-                                                #[cfg(all(feature = "prometheus", not(test)))]
-                                                {
-                                                    METRIC_RECEIVED_ACKS.increment(&["true"]);
-                                                    METRIC_TICKETS_COUNT.increment(&["losing"]);
-                                                }
+                                                METRIC_RECEIVED_ACKS.increment(&["true"]);
+                                                METRIC_TICKETS_COUNT.increment(&["losing"]);
                                             }
                                         }
 
                                         Some(AckProcessed::Receive(peer, Ok(reply)))
                                     }
                                     Err(e) => {
-                                        error!(
-                                            "Encountered error while handling acknowledgement from peer '{}': {}",
-                                            &peer, e
-                                        );
-
                                         #[cfg(all(feature = "prometheus", not(test)))]
                                         METRIC_RECEIVED_ACKS.increment(&["false"]);
+
+                                        error!(
+                                            "Encountered error while handling acknowledgement from peer '{peer}': {e}",
+                                        );
 
                                         None
                                     }
